Configure Security (#12)
This pull request adds a new security workflow to the repository that
includes security checks using `bundler-audit` and `brakeman` gems.

The security workflow runs on every push to the repository and checks
for known vulnerabilities in dependencies and performs static analysis
security vulnerability scanning for Ruby on Rails applications.

The workflow runs on an Ubuntu latest operating system and uses the
latest version of Ruby. The `bundler-audit` and `brakeman` gems are
added to the Gemfile and are installed along with the other gems when
the workflow runs.

With the addition of this security workflow, developers can ensure that
the codebase is free of known security vulnerabilities and maintain the
overall security of the application.