Configure RSpec (#13)
This pull request introduces a new RSpec workflow to the repository that
automates the testing process for the application using GitHub Actions.

The RSpec workflow is triggered on every push to the repository and runs
the test suite on an Ubuntu latest operating system. It sets up the
required Ruby environment, installs the necessary gems, and sets up
PostgreSQL as a service with the required configuration.

In addition to the RSpec workflow, this pull request updates the
`loofah` gem to version 2.20.0 and the `tailwindcss-rails` gem to
version 2.0.27, ensuring compatibility and the latest features.

With the addition of this RSpec workflow, developers can now
automatically run the test suite, ensuring that new code changes do not
introduce regressions and maintain the overall quality of the
application.